[DCS-204] add spark integration support
