+++ examples/configurations/elasticsearch/data_source.yaml
@@ -1,0 +1,9 @@
+# Data sources to query
+data_sources:
+  - name: iris_elasticsearch
+    type: elasticsearch
+    connection:
+      host: 127.0.0.1
+      port: 9207
+      username: !ENV ${OS_USER}
+      password: !ENV ${OS_PASS}
