+++ examples/configurations/mysql/example_mysql_config.yaml
@@ -1,0 +1,3 @@
+validations for iris_mysql.dcs_iris:
+  - row count:
+      on: count_rows
