--- examples/configurations/old/example_bigquery_config.yaml
@@ -1,146 +1,0 @@
-# Data sources to query
-data_sources:
-  - name: iris_bigquery
-    type: bigquery
-    connection:
-      project: !ENV ${GOOGLE_PROJECT_ID}
-      dataset: !ENV ${GOOGLE_BIGQUERY_DATASET}
-      credentials_base64: !ENV ${GOOGLE_APPLICATION_CREDENTIALS_BASE64}
-metrics:
-  # Numeric Metrics - bigquery
-  - name: bigquery_avg_price
-    metric_type: avg
-    resource: iris_bigquery.dcs_iris.sepal_length
-
-  - name: bigquery_min_price
-    metric_type: min
-    resource: iris_bigquery.dcs_iris.sepal_length
-
-  - name: bigquery_max_price
-    metric_type: max
-    resource: iris_bigquery.dcs_iris.sepal_length
-
-  - name: bigquery_variance_price
-    metric_type: variance
-    resource: iris_bigquery.dcs_iris.sepal_length
-
-  # Uniqueness Metrics - bigquery
-  - name: bigquery_distinct_count_price
-    metric_type: distinct_count
-    resource: iris_bigquery.dcs_iris.sepal_length
-
-  - name: bigquery_duplicate_count_price
-    metric_type: duplicate_count
-    resource: iris_bigquery.dcs_iris.sepal_length
-
-  # Completeness Metrics - bigquery
-  - name: bigquery_null_count_price
-    metric_type: null_count
-    resource: iris_bigquery.dcs_iris.sepal_length
-
-  - name: bigquery_empty_string_count_price
-    metric_type: empty_string_count
-    resource: iris_bigquery.dcs_iris.species
-
-  # Reliability Metrics - bigquery
-  - name: bigquery_row_count
-    metric_type: row_count
-    resource: iris_bigquery.dcs_iris
-
-  - name: bigquery_freshness
-    metric_type: freshness
-    resource: iris_bigquery.dcs_iris.timestamp
-
-  # Combined Metrics - bigquery
-  - name: bigquery_combined_div
-    metric_type: combined
-    expression: div(bigquery_distinct_count_price, bigquery_max_price)
-
-  - name: bigquery_combined_mul
-    metric_type: combined
-    expression: mul(bigquery_distinct_count_price, bigquery_max_price)
-
-  - name: bigquery_combined_sum
-    metric_type: combined
-    expression: sum(bigquery_distinct_count_price, bigquery_max_price)
-
-  - name: bigquery_combined_sub
-    metric_type: combined
-    expression: sub(bigquery_distinct_count_price, bigquery_max_price)
-
-  - name: bigquery_combined_div_mul
-    metric_type: combined
-    expression: div(mul(bigquery_distinct_count_price, bigquery_max_price), bigquery_min_price)
-
-  - name: bigquery_combined_div_mul_sum
-    metric_type: combined
-    expression: sum(div(mul(bigquery_distinct_count_price, bigquery_max_price), bigquery_min_price), bigquery_row_count)
-
-  - name: bigquery_combined_div_mul_sum_sub
-    metric_type: combined
-    expression: sub(sum(div(mul(bigquery_distinct_count_price, bigquery_max_price), bigquery_min_price), bigquery_row_count), bigquery_duplicate_count_price)
-
-  - name: bigquery_combined_div_mul_sum_sub_div
-    metric_type: combined
-    expression: div(sub(sum(div(mul(bigquery_distinct_count_price, bigquery_max_price), bigquery_min_price), bigquery_row_count), bigquery_duplicate_count_price), bigquery_distinct_count_price)
-
-  # With Filters
-  - name: bigquery_avg_price_with_filter
-    metric_type: avg
-    resource: iris_bigquery.dcs_iris.sepal_length
-    filters:
-      where: 'species = "setosa"'
-
-  - name: bigquery_max_price_with_filter
-    metric_type: max
-    resource: iris_bigquery.dcs_iris.sepal_length
-    filters:
-      where: 'species = "setosa"'
-
-  - name: bigquery_min_price_with_filter
-    metric_type: min
-    resource: iris_bigquery.dcs_iris.sepal_length
-    filters:
-      where: 'species = "setosa"'
-
-  - name: bigquery_variance_price_with_filter
-    metric_type: variance
-    resource: iris_bigquery.dcs_iris.sepal_length
-    filters:
-      where: 'species = "setosa"'
-
-  - name: bigquery_distinct_count_price_with_filter
-    metric_type: distinct_count
-    resource: iris_bigquery.dcs_iris.sepal_length
-    filters:
-      where: 'species = "setosa"'
-
-  - name: bigquery_duplicate_count_price_with_filter
-    metric_type: duplicate_count
-    resource: iris_bigquery.dcs_iris.sepal_length
-    filters:
-      where: 'species = "setosa"'
-
-  - name: bigquery_null_count_price_with_filter
-    metric_type: null_count
-    resource: iris_bigquery.dcs_iris.sepal_length
-    filters:
-      where: 'species = "setosa"'
-
-  - name: bigquery_empty_string_count_price_with_filter
-    metric_type: empty_string_count
-    resource: iris_bigquery.dcs_iris.species
-    filters:
-      where: 'species = "setosa"'
-
-  - name: bigquery_row_count_with_filter
-    metric_type: row_count
-    resource: iris_bigquery.dcs_iris
-    filters:
-      where: 'species = "setosa"'
-
-  - name: bigquery_freshness_with_filter
-    metric_type: freshness
-    resource: iris_bigquery.dcs_iris.timestamp
-    filters:
-      where: 'species = "setosa"'
