--- examples/configurations/old/example_databricks_config.yaml
@@ -1,135 +1,0 @@
-# Data sources to query
-data_sources:
-  - name: iris_databricks
-    type: databricks
-    connection:
-      host: !ENV ${DATABRICKS_HOST}
-      token: !ENV ${DATABRICKS_TOKEN}
-      http_path: !ENV ${DATABRICKS_HTTP_PATH}
-      catalog: test_dcs
-      schema: test_schema
-
-metrics:
-  # Numeric Metrics - Databricks
-  - name: databricks_avg_price
-    metric_type: avg
-    resource: iris_databricks.dcs_iris.sepal_length
-
-  - name: databricks_min_price
-    metric_type: min
-    resource: iris_databricks.dcs_iris.sepal_length
-
-  - name: databricks_max_price
-    metric_type: max
-    resource: iris_databricks.dcs_iris.sepal_length
-
-  - name: databricks_variance_price
-    metric_type: variance
-    resource: iris_databricks.dcs_iris.sepal_length
-
-  # Uniqueness Metrics - Databricks
-  - name: databricks_distinct_count_price
-    metric_type: distinct_count
-    resource: iris_databricks.dcs_iris.sepal_length
-
-  - name: databricks_duplicate_count_price
-    metric_type: duplicate_count
-    resource: iris_databricks.dcs_iris.sepal_length
-
-  # Completeness Metrics - Databricks
-  - name: databricks_null_count_price
-    metric_type: null_count
-    resource: iris_databricks.dcs_iris.sepal_length
-
-  - name: databricks_null_percentage_price
-    metric_type: null_percentage
-    resource: iris_databricks.dcs_iris.sepal_length
-
-  - name: databricks_empty_string_count_price
-    metric_type: empty_string_count
-    resource: iris_databricks.dcs_iris.species
-
-  # Reliability Metrics - Databricks
-  - name: databricks_row_count
-    metric_type: row_count
-    resource: iris_databricks.dcs_iris
-
-  - name: databricks_freshness
-    metric_type: freshness
-    resource: iris_databricks.dcs_iris.timestamp
-
-  # Combined Metrics - Databricks
-  - name: databricks_combined_div
-    metric_type: combined
-    expression: div(databricks_distinct_count_price, databricks_max_price)
-
-  - name: databricks_combined_mul
-    metric_type: combined
-    expression: mul(databricks_distinct_count_price, databricks_max_price)
-
-  - name: databricks_combined_sum
-    metric_type: combined
-    expression: sum(databricks_distinct_count_price, databricks_max_price)
-
-  - name: databricks_combined_sub
-    metric_type: combined
-    expression: sub(databricks_distinct_count_price, databricks_max_price)
-
-  - name: databricks_combined_div_mul
-    metric_type: combined
-    expression: div(mul(databricks_distinct_count_price, databricks_max_price), databricks_min_price)
-
-  - name: databricks_combined_div_mul_sum
-    metric_type: combined
-    expression: sum(div(mul(databricks_distinct_count_price, databricks_max_price), databricks_min_price), databricks_document_count)
-
-  - name: databricks_combined_div_mul_sum_sub
-    metric_type: combined
-    expression: sub(sum(div(mul(databricks_distinct_count_price, databricks_max_price), databricks_min_price), databricks_document_count), databricks_duplicate_count_price)
-
-  - name: databricks_combined_div_mul_sum_sub_div
-    metric_type: combined
-    expression: div(sub(sum(div(mul(databricks_distinct_count_price, databricks_max_price), databricks_min_price), databricks_document_count), databricks_duplicate_count_price), databricks_distinct_count_price)
-
-  # With Filters
-  - name: databricks_avg_price_with_filter
-    metric_type: avg
-    resource: iris_databricks.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: databricks_avg_price_with_filter
-    metric_type: avg
-    resource: iris_databricks.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: databricks_max_price_with_filter
-    metric_type: max
-    resource: iris_databricks.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: databricks_min_price_with_filter
-    metric_type: min
-    resource: iris_databricks.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: databricks_variance_price_with_filter
-    metric_type: variance
-    resource: iris_databricks.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: databricks_distinct_count_price_with_filter
-    metric_type: distinct_count
-    resource: iris_databricks.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: databricks_duplicate_count_price_with_filter
-    metric_type: duplicate_count
-    resource: iris_databricks.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
