--- examples/configurations/old/example_elasticsearch_config.yaml
@@ -1,134 +1,0 @@
-# Data sources to query
-data_sources:
-  - name: iris_elasticsearch
-    type: elasticsearch
-    connection:
-      host: 127.0.0.1
-      port: 9207
-      username: !ENV ${OS_USER}
-      password: !ENV ${OS_PASS}
-
-metrics:
-  # Numeric Metrics - ElasticSearch
-  - name: elasticsearch_avg_price
-    metric_type: avg
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-
-  - name: elasticsearch_min_price
-    metric_type: min
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-
-  - name: elasticsearch_max_price
-    metric_type: max
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-
-  - name: elasticsearch_variance_price
-    metric_type: variance
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-
-  # Uniqueness Metrics - ElasticSearch
-  - name: elasticsearch_distinct_count_price
-    metric_type: distinct_count
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-
-  - name: elasticsearch_duplicate_count_price
-    metric_type: duplicate_count
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-
-  # Completeness Metrics - ElasticSearch
-  - name: elasticsearch_null_count_price
-    metric_type: null_count
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-
-  - name: elasticsearch_null_percentage_price
-    metric_type: null_percentage
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-
-  - name: elasticsearch_empty_string_count_price
-    metric_type: empty_string_count
-    resource: iris_elasticsearch.dcs_iris.species
-
-  # Reliability Metrics - ElasticSearch
-  - name: elasticsearch_document_count
-    metric_type: document_count
-    resource: iris_elasticsearch.dcs_iris
-
-  - name: elasticsearch_freshness
-    metric_type: freshness
-    resource: iris_elasticsearch.dcs_iris.timestamp
-
-  # Combined Metrics - ElasticSearch
-  - name: elasticsearch_combined_div
-    metric_type: combined
-    expression: div(elasticsearch_distinct_count_price, elasticsearch_max_price)
-
-  - name: elasticsearch_combined_mul
-    metric_type: combined
-    expression: mul(elasticsearch_distinct_count_price, elasticsearch_max_price)
-
-  - name: elasticsearch_combined_sum
-    metric_type: combined
-    expression: sum(elasticsearch_distinct_count_price, elasticsearch_max_price)
-
-  - name: elasticsearch_combined_sub
-    metric_type: combined
-    expression: sub(elasticsearch_distinct_count_price, elasticsearch_max_price)
-
-  - name: elasticsearch_combined_div_mul
-    metric_type: combined
-    expression: div(mul(elasticsearch_distinct_count_price, elasticsearch_max_price), elasticsearch_min_price)
-
-  - name: elasticsearch_combined_div_mul_sum
-    metric_type: combined
-    expression: sum(div(mul(elasticsearch_distinct_count_price, elasticsearch_max_price), elasticsearch_min_price), elasticsearch_document_count)
-
-  - name: elasticsearch_combined_div_mul_sum_sub
-    metric_type: combined
-    expression: sub(sum(div(mul(elasticsearch_distinct_count_price, elasticsearch_max_price), elasticsearch_min_price), elasticsearch_document_count), elasticsearch_duplicate_count_price)
-
-  - name: elasticsearch_combined_div_mul_sum_sub_div
-    metric_type: combined
-    expression: div(sub(sum(div(mul(elasticsearch_distinct_count_price, elasticsearch_max_price), elasticsearch_min_price), elasticsearch_document_count), elasticsearch_duplicate_count_price), elasticsearch_distinct_count_price)
-
-  # With Filters
-
-  - name: elasticsearch_max_price_with_filter
-    metric_type: max
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: elasticsearch_min_price_with_filter
-    metric_type: min
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: elasticsearch_avg_price_with_filter
-    metric_type: avg
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: elasticsearch_max_price_with_filter
-    metric_type: max
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: elasticsearch_min_price_with_filter
-    metric_type: min
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-  - name: elasticsearch_variance_price_with_filter
-    metric_type: variance
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: elasticsearch_distinct_count_price_with_filter
-    metric_type: distinct_count
-    resource: iris_elasticsearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
