--- examples/configurations/old/example_mysql_config.yaml
@@ -1,128 +1,0 @@
-# Data sources to query
-data_sources:
-  - name: iris_mysql
-    type: mysql
-    connection:
-      host: 127.0.0.1
-      port: 3306
-      username: !ENV ${MYSQL_USER}
-      password: !ENV ${MYSQL_PASS}
-      database: dcs_db
-metrics:
-  # Numeric Metrics
-  - name: mysql_avg_price
-    metric_type: avg
-    resource: iris_mysql.dcs_iris.sepal_length
-
-  - name: mysql_min_price
-    metric_type: min
-    resource: iris_mysql.dcs_iris.sepal_length
-
-  - name: mysql_max_price
-    metric_type: max
-    resource: iris_mysql.dcs_iris.sepal_length
-
-  - name: mysql_variance_price
-    metric_type: variance
-    resource: iris_mysql.dcs_iris.sepal_length
-
-  # Uniqueness Metrics
-  - name: mysql_distinct_count_price
-    metric_type: distinct_count
-    resource: iris_mysql.dcs_iris.sepal_length
-
-  - name: mysql_duplicate_count_price
-    metric_type: duplicate_count
-    resource: iris_mysql.dcs_iris.sepal_length
-
-  # Completeness Metrics
-  - name: mysql_null_count_price
-    metric_type: null_count
-    resource: iris_mysql.dcs_iris.sepal_length
-
-  - name: mysql_null_percentage_price
-    metric_type: null_percentage
-    resource: iris_mysql.dcs_iris.sepal_length
-
-  - name: mysql_empty_string_count_price
-    metric_type: empty_string_count
-    resource: iris_mysql.dcs_iris.species
-
-  # Reliability Metrics
-  - name: mysql_row_count
-    metric_type: row_count
-    resource: iris_mysql.dcs_iris
-
-  - name: mysql_freshness
-    metric_type: freshness
-    resource: iris_mysql.dcs_iris.timestamp
-
-  # Combined Metrics
-  - name: mysql_combined_div
-    metric_type: combined
-    expression: div(mysql_distinct_count_price, mysql_max_price)
-
-  - name: mysql_combined_mul
-    metric_type: combined
-    expression: mul(mysql_distinct_count_price, mysql_max_price)
-
-  - name: mysql_combined_sum
-    metric_type: combined
-    expression: sum(mysql_distinct_count_price, mysql_max_price)
-
-  - name: mysql_combined_sub
-    metric_type: combined
-    expression: sub(mysql_distinct_count_price, mysql_max_price)
-
-  - name: mysql_combined_div_mul
-    metric_type: combined
-    expression: div(mul(mysql_distinct_count_price, mysql_max_price), mysql_min_price)
-
-  - name: mysql_combined_div_mul_sum
-    metric_type: combined
-    expression: sum(div(mul(mysql_distinct_count_price, mysql_max_price), mysql_min_price), mysql_row_count)
-
-  - name: mysql_combined_div_mul_sum_sub
-    metric_type: combined
-    expression: sub(sum(div(mul(mysql_distinct_count_price, mysql_max_price), mysql_min_price), mysql_row_count), mysql_duplicate_count_price)
-
-  - name: mysql_combined_div_mul_sum_sub_div
-    metric_type: combined
-    expression: div(sub(sum(div(mul(mysql_distinct_count_price, mysql_max_price), mysql_min_price), mysql_row_count), mysql_duplicate_count_price), mysql_distinct_count_price)
-
-  # Filtered Metrics
-  - name: mysql_max_price_filtered
-    metric_type: max
-    resource: iris_mysql.dcs_iris.sepal_length
-    filters:
-      where: "species = 'versicolor'"
-
-  - name: mysql_avg_price_with_filter
-    metric_type: avg
-    resource: iris_mysql.dcs_iris.sepal_length
-    filters:
-      where: "species = 'versicolor'"
-
-  - name: mysql_max_price_with_filter
-    metric_type: max
-    resource: iris_mysql.dcs_iris.sepal_length
-    filters:
-      where: "species = 'versicolor'"
-
-  - name: mysql_min_price_with_filter
-    metric_type: min
-    resource: iris_mysql.dcs_iris.sepal_length
-    filters:
-      where: "species = 'versicolor'"
-
-  - name: mysql_distinct_count_price_with_filter
-    metric_type: distinct_count
-    resource: iris_mysql.dcs_iris.sepal_length
-    filters:
-      where: "species = 'versicolor'"
-
-  - name: mysql_empty_string_count_price_with_filter
-    metric_type: empty_string_count
-    resource: iris_mysql.dcs_iris.species
-    filters:
-      where: "species = 'versicolor'"
