--- examples/configurations/old/example_opensearch_config.yaml
@@ -1,134 +1,0 @@
-# Data sources to query
-data_sources:
-  - name: iris_opensearch
-    type: opensearch
-    connection:
-      host: 127.0.0.1
-      port: 9205
-      username: !ENV ${OS_USER}
-      password: !ENV ${OS_PASS}
-
-metrics:
-  # Numeric Metrics - OpenSearch
-  - name: opensearch_avg_price
-    metric_type: avg
-    resource: iris_opensearch.dcs_iris.sepal_length
-
-  - name: opensearch_min_price
-    metric_type: min
-    resource: iris_opensearch.dcs_iris.sepal_length
-
-  - name: opensearch_max_price
-    metric_type: max
-    resource: iris_opensearch.dcs_iris.sepal_length
-
-  - name: opensearch_variance_price
-    metric_type: variance
-    resource: iris_opensearch.dcs_iris.sepal_length
-
-  # Uniqueness Metrics - OpenSearch
-  - name: opensearch_distinct_count_price
-    metric_type: distinct_count
-    resource: iris_opensearch.dcs_iris.sepal_length
-
-  - name: opensearch_duplicate_count_price
-    metric_type: duplicate_count
-    resource: iris_opensearch.dcs_iris.sepal_length
-
-  # Completeness Metrics - OpenSearch
-  - name: opensearch_null_count_price
-    metric_type: null_count
-    resource: iris_opensearch.dcs_iris.sepal_length
-
-  - name: opensearch_null_percentage_price
-    metric_type: null_percentage
-    resource: iris_opensearch.dcs_iris.sepal_length
-
-  - name: opensearch_empty_string_count_price
-    metric_type: empty_string_count
-    resource: iris_opensearch.dcs_iris.species
-
-  # Reliability Metrics - OpenSearch
-  - name: opensearch_document_count
-    metric_type: document_count
-    resource: iris_opensearch.dcs_iris
-
-  - name: opensearch_freshness
-    metric_type: freshness
-    resource: iris_opensearch.dcs_iris.timestamp
-
-  # Combined Metrics - OpenSearch
-  - name: opensearch_combined_div
-    metric_type: combined
-    expression: div(opensearch_distinct_count_price, opensearch_max_price)
-
-  - name: opensearch_combined_mul
-    metric_type: combined
-    expression: mul(opensearch_distinct_count_price, opensearch_max_price)
-
-  - name: opensearch_combined_sum
-    metric_type: combined
-    expression: sum(opensearch_distinct_count_price, opensearch_max_price)
-
-  - name: opensearch_combined_sub
-    metric_type: combined
-    expression: sub(opensearch_distinct_count_price, opensearch_max_price)
-
-  - name: opensearch_combined_div_mul
-    metric_type: combined
-    expression: div(mul(opensearch_distinct_count_price, opensearch_max_price), opensearch_min_price)
-
-  - name: opensearch_combined_div_mul_sum
-    metric_type: combined
-    expression: sum(div(mul(opensearch_distinct_count_price, opensearch_max_price), opensearch_min_price), opensearch_document_count)
-
-  - name: opensearch_combined_div_mul_sum_sub
-    metric_type: combined
-    expression: sub(sum(div(mul(opensearch_distinct_count_price, opensearch_max_price), opensearch_min_price), opensearch_document_count), opensearch_duplicate_count_price)
-
-  - name: opensearch_combined_div_mul_sum_sub_div
-    metric_type: combined
-    expression: div(sub(sum(div(mul(opensearch_distinct_count_price, opensearch_max_price), opensearch_min_price), opensearch_document_count), opensearch_duplicate_count_price), opensearch_distinct_count_price)
-
-  # With Filters
-  - name: opensearch_avg_price_with_filter
-    metric_type: avg
-    resource: iris_opensearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: opensearch_max_price_with_filter
-    metric_type: max
-    resource: iris_opensearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: opensearch_min_price_with_filter
-    metric_type: min
-    resource: iris_opensearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: opensearch_distinct_count_price_with_filter
-    metric_type: distinct_count
-    resource: iris_opensearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: opensearch_duplicate_count_price_with_filter
-    metric_type: duplicate_count
-    resource: iris_opensearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: opensearch_null_count_price_with_filter
-    metric_type: null_count
-    resource: iris_opensearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
-
-  - name: opensearch_null_percentage_price_with_filter
-    metric_type: null_percentage
-    resource: iris_opensearch.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
