--- examples/configurations/old/example_redshift_config.yaml
@@ -1,104 +1,0 @@
-# Data sources to query
-data_sources:
-  - name: iris_redshift
-    type: redshift
-    connection:
-      host: !ENV ${REDSHIFT_HOST}
-      port: !ENV ${REDSHIFT_PORT}
-      username: !ENV ${REDSHIFT_USERNAME}
-      password: !ENV ${REDSHIFT_PASSWORD}
-      database: !ENV ${REDSHIFT_DATABASE}
-
-metrics:
-  # Numeric Metrics - RedShift
-  - name: redshift_avg_price
-    metric_type: avg
-    resource: iris_redshift.dcs_iris.sepal_length
-
-  - name: redshift_min_price
-    metric_type: min
-    resource: iris_redshift.dcs_iris.sepal_length
-
-  - name: redshift_max_price
-    metric_type: max
-    resource: iris_redshift.dcs_iris.sepal_length
-
-  - name: redshift_variance_price
-    metric_type: variance
-    resource: iris_redshift.dcs_iris.sepal_length
-
-  # Uniqueness Metrics - RedShift
-  - name: redshift_distinct_count_price
-    metric_type: distinct_count
-    resource: iris_redshift.dcs_iris.sepal_length
-    validation:
-      threshold: "<= 36"
-
-  - name: redshift_duplicate_count_price
-    metric_type: duplicate_count
-    resource: iris_redshift.dcs_iris.sepal_length
-    validation:
-      threshold: ">= 9 & >= 36"
-
-  # Completeness Metrics - RedShift
-  - name: redshift_null_count_price
-    metric_type: null_count
-    resource: iris_redshift.dcs_iris.sepal_length
-
-  - name: redshift_null_percentage_price
-    metric_type: null_percentage
-    resource: iris_redshift.dcs_iris.sepal_length
-
-  - name: redshift_empty_string_count_price
-    metric_type: empty_string_count
-    resource: iris_redshift.dcs_iris.species
-
-  # Reliability Metrics - RedShift
-  - name: redshift_row_count
-    metric_type: row_count
-    resource: iris_redshift.dcs_iris
-
-  # Combined Metrics - RedShift
-  - name: redshift_combined_div
-    metric_type: combined
-    expression: div(redshift_distinct_count_price, redshift_max_price)
-
-  - name: redshift_combined_mul
-    metric_type: combined
-    expression: mul(redshift_distinct_count_price, redshift_max_price)
-
-  - name: redshift_combined_sum
-    metric_type: combined
-    expression: sum(redshift_distinct_count_price, redshift_max_price)
-
-  - name: redshift_combined_sub
-    metric_type: combined
-    expression: sub(redshift_distinct_count_price, redshift_max_price)
-
-  # - name: redshift_combined_percentage
-  #   metric_type: combined
-  #   expression: percentage(redshift_distinct_count_price, redshift_max_price)
-
-  - name: redshift_combined_div_mul
-    metric_type: combined
-    expression: div(mul(redshift_distinct_count_price, redshift_max_price), redshift_min_price)
-
-  - name: redshift_combined_div_mul_sum
-    metric_type: combined
-    expression: sum(div(mul(redshift_distinct_count_price, redshift_max_price), redshift_min_price), redshift_row_count)
-
-  - name: redshift_combined_div_mul_sum_sub
-    metric_type: combined
-    expression: sub(sum(div(mul(redshift_distinct_count_price, redshift_max_price), redshift_min_price), redshift_row_count), redshift_duplicate_count_price)
-
-  - name: redshift_combined_div_mul_sum_sub_div
-    metric_type: combined
-    expression: div(sub(sum(div(mul(redshift_distinct_count_price, redshift_max_price), redshift_min_price), redshift_row_count), redshift_duplicate_count_price), redshift_distinct_count_price)
-    validation:
-      threshold: "<= 0.5"
-  # With Filters
-  - name: redshift_avg_price_with_filter
-    metric_type: avg
-    resource: iris_redshift.dcs_iris.sepal_length
-    filters:
-      where: '{ "match": { "species": "versicolor" } }'
