+++ examples/configurations/postgres/example_postgres_config.yaml
@@ -1,15 +1,3 @@
-# Data sources to query
-data_sources:
-  - name: iris_pgsql
-    type: postgres
-    connection:
-      host: 127.0.0.1
-      port: 5421
-      username: !ENV ${PGSQL_USER}
-      password: !ENV ${PGSQL_PASS}
-      database: dcs_db
-      schema: public
-
 validations for iris_pgsql.dcs_iris:
   # Numeric Metrics
   - sepal length min size:
