+++ pyproject.toml
@@ -42,7 +42,7 @@ pymysql = { version="^1.1.0", optional=true, extras=["rsa"] }
 snowflake-sqlalchemy = { version="^1.5.3", optional=true }
 pyodbc = { version="^5.0.1", optional=true }
 psycopg2-binary = {version = "^2.9.9", optional = true}
-
+pyspark = { version="^3.2.1", optional=true }
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.1.3"
@@ -53,6 +53,9 @@ pre-commit = "^2.20.0"
 pytest-cov = "^4.0.0"
 python-dotenv = "^1.0.0"
 pandas = "^1.4.0"
+jupyter = "^1.0.0"
+pyspark = "^3.2.1"
+findspark = "^1.4.2"
 
 [tool.poetry.group.docs]
 optional = true
@@ -71,6 +74,7 @@ bigquery = ["sqlalchemy-bigquery"]
 databricks = ["databricks-sql-connector"]
 elasticsearch = ["elasticsearch"]
 snowflake = ["snowflake-sqlalchemy"]
+spark = ["pyspark"]
 all = [
     "psycopg2-binary",
     "pymysql",
@@ -79,7 +83,8 @@ all = [
     "databricks-sql-connector",
     "elasticsearch",
     "snowflake-sqlalchemy",
-    "pyodbc"
+    "pyodbc",
+    "pyspark"
 ]
 
 [tool.poetry.scripts]
